Move arrow/util/concatenate* to arrow/array
