+++ cpp/src/arrow/array/concatenate-test.cc
@@ -30,13 +30,13 @@
 #include <gtest/gtest.h>
 
 #include "arrow/array.h"
+#include "arrow/array/concatenate.h"
 #include "arrow/buffer.h"
 #include "arrow/status.h"
 #include "arrow/testing/gtest_common.h"
 #include "arrow/testing/random.h"
 #include "arrow/testing/util.h"
 #include "arrow/type.h"
-#include "arrow/util/concatenate.h"
 
 namespace arrow {
