+++ cpp/src/arrow/array/concatenate.h
