+++ cpp/src/arrow/util/CMakeLists.txt
@@ -29,7 +29,6 @@ arrow_install_all_headers("arrow/util")
 add_arrow_test(bit-util-test)
 add_arrow_test(checked-cast-test)
 add_arrow_test(compression-test)
-add_arrow_test(concatenate-test)
 add_arrow_test(decimal-test)
 add_arrow_test(hashing-test)
 add_arrow_test(int-util-test)
